refactor: consolidate library store
