+++ frontend/src/store/libraryStore.js
@@ -1,22 +1,27 @@
 import { create } from "zustand";
+import { persist } from "zustand/middleware";
 import { fetchLibrary as apiFetchLibrary } from "@/services/libraryService";
 
-const useLibraryStore = create((set) => ({
-  items: [],
-  loading: false,
-  fetchLibrary: async () => {
-    set({ loading: true });
-    try {
-      const items = await apiFetchLibrary();
-      const activeItems = Array.isArray(items)
-        ? items.filter((b) => b?.status === "active")
-        : [];
-      set({ items: activeItems, loading: false });
-    } catch (err) {
-      set({ loading: false });
-    }
-  },
-}));
+const useLibraryStore = create(
+  persist(
+    (set) => ({
+      books: [],
+      loading: false,
+      error: null,
+      fetchLibrary: async () => {
+        set({ loading: true, error: null });
+        try {
+          const data = await apiFetchLibrary();
+          set({ books: Array.isArray(data) ? data : [], loading: false });
+        } catch (err) {
+          set({ error: err.message, loading: false });
+        }
+      },
+      clear: () => set({ books: [] }),
+    }),
+    { name: "library-store" }
+  )
+);
 
 export default useLibraryStore;
